requirements.txt
aiogram==2.19
aiogram-dialog==1.4.1
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
hiredis==2.0.0
idna==3.3
Jinja2==3.1.1
magic-filter==1.0.6
MarkupSafe==2.1.1
motor==2.5.1
multidict==6.0.2
pymongo==3.12.3
python-dotenv==0.20.0
pytz==2022.1
typing_extensions==4.1.1
ujson==5.1.0
yarl==1.7.2
